"Refactor comment panel creation logic"
The diff covers a wide range of changes primarily focused on refactoring the logic of creating comment panels in the GUI. This includes modifications in multiple classes. In general, the changes help to simplify the code structure by removing unnecessary codes, fixing todos and fixmes, as well as making the code more organized and easily readable. For instance, in CustomContentContainer class, an event listener has been added to dynamically adjust the top panel’s dimensions based on the size of the content container. In CustomCommentPanel class, the Git diff shows an improvement in how chat bubbles are created, tilted, and displayed in the left and right side of the GUI. Similarly, in CommentManagerImpl, a method is created to check if the client matches the sender of the message. The update in CustomWebsocketClient is more about enhancing the error handling where the original exception is now directly thrown as RuntimeException. In ChatPanel, the size of comment window is adjusted from 700x700 to 750x700. Improvements have also been made to handle the comments and replies based on different message types using the CommentTypeEnum and BaseModel classes. The JSON classes also had minor changes to include an id field for deduplication purposes.