"Refactor CustomQuotePanel for BaseModel use and added Excalidraw file"
Refactored CustomQuotePanel class to use a more generic BaseModel instance instead of the more specific MessageModel. This was done to allow the CustomQuotePanel to accept other types of models in the future. Also, some corrections and adjustments were made to handle BaseModel properly within createQuotedTextPane() method. In addition, an excalidraw file was added to content. The Excalidraw diagram details the "reply_panel" layout structure and its interactivity.