"Enhance code readability and functionality in comment GUI"
Refactored and enhanced code of classes handling comments in the GUI. Renamed CustomQuotePanel to CustomReferencePanel to clarify purpose and generalised to BaseModel for more inclusive handling of comments. Reorganised layout for better visibility and readability of comments and ease of interaction with link and reply comments. Added the setupNameAndTimeTopPanel and populateReferencePanel methods for customization. This improves maintainability and readability of the code, while improving user experience.