"Integrate 'MainFrameGuiInterface' into GUI comment feature"
In this update, 'MainFrameGuiInterface' was integrated into the comment feature GUI classes which includes 'CustomReferencePanel', 'CustomTextAndQuoteForBubblePanel', 'CustomQuoteBubblePanel', 'TextMessageFactory', 'CustomLinkPanel', 'CustomLinkCommentTextPane', 'CustomCommentPanel', 'LinkPanelFactory', 'ReferencePanelFactory', 'PicturePanelFactory', 'CustomSimpleTextPane', 'CustomReplyPanel', 'CustomTextPane', and others. A number of constructors and methods were adjusted for improved data flow and accessibility, ensuring a more efficient and responsive interface for users. This change brings better interactivity of the Graphical User Interface of Comments and improves maintainability of the codebase.