"Refactor CustomPictureLabel and related classes"
Removed the ComponentListener interface from the CustomPictureLabel and cleaned up the redundant code. Also, eliminated the unnecessary method addPictureAsIconToLabel() from the instance of CustomPictureLabel in PicturePanelFactory, and initiated scaling, if necessary, is taken care by the PictureSwingWorker during construction. This refactoring improves code readability and reduces clutter. Also, removed code that tried to handle picture resizing themselves and offloaded these tasks to the classes which are designed to handle this, ensuring Single Responsibility Principle is adhered to. In CustomPictureMaximizeMenuItem, moved the task of opening the image buffer in an external viewer to a virtual thread for better responsiveness.