"Introduced clickable links and copy functionality"
Add mouse event handlers to CustomLinkTextPane to allow opening of links directly from the GUI. Implemented the mouse listener interface and added methods to handle mouse events for this. This effectively allows users to click on a link displayed in the panel and have the link open in their default browser.

Also, a new file, CopyTextMenuItem.java was created which implements a custom JMenuItem that allows for copying text displayed in a CustomSimpleTextPane By incorporating an actionListener, the JMenuItem copies the selected text, or if there's no selection, the entire text from a CustomSimpleTextPane and puts it into the system clipboard, giving the users ability to easily copy text from the panels.

Lastly, changes to the CustomSimpleTextPane were made where mouse event methods were added so right clicking will now open a context menu with copy functionality.

These changes will increase end user interaction by allowing them to interact with the links and text more directly.