Adjust ReactionPopupMenuImpl to use topPanel instead of layeredPane
The recent changes were made to refactor the ReactionPopupMenuImpl and ReactionPopupHandler classes to use a JPanel (topPanel) instead of a JLayeredPane (layeredPane). This refactor brings simplification of the GUI hierarchy and reduces the complexity of the layeredPane components. Moreover, defensive coding was implemented in the BaseModel class to initialize the userInteractions list if it's null. Also, modifications to TransparentTopPanel were made to handle the popup menu animations. To achieve a more seamless visual experience, the border color is now only changed when a mouse hovers over it.